requirements/requirements.txt
Django==2.2.9
dj-database-url==0.5.0
django-extensions==2.1.3
django-filetransfers==0.1.0
django-filter<3.0.0,>=2.2.0
django-model-utils==3.1.2
django-crispy-forms==1.7.2
django-currentuser==0.4.2
django-guardian==1.4.9
django-fontawesome==1.0
djangorestframework==3.11.0
celery==4.3.0
django-celery-beat==1.5.0
#rules==2.0
wagtail==2.8rc1
PyJWT==1.6.4
psycopg2==2.8.3
graphene<3.0.0,>=2.1.8
graphene-django>=2.8.2
graphene-django-extras>=0.4.8
graphene-file-upload==1.2.2
graphql-core<3,>=2.2
graphql-relay<3,>=2
django-graphql-extensions>=0.0.7
idna==2.6
Inflector==2.0.12
jsonfield==2.0.2
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
pinax-messages==2.0.2
pinax-templates==2.0.2
pinax-notifications==5.0.3
promise<3,>=2.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
flake8==3.5.0
python-codicefiscale==0.3.4
django-bootstrap-form